Refactor: Centralize security validation in perl-dap
Moved safe evaluation logic and regexes from DebugAdapter to security module.
Unified validation logic for expressions and breakpoint conditions.
Fixed flaky security tests by using tempfile for isolation.
